requirements.txt
Flask
Flask-Cors
Flask_RESTful
pymongo
PyYAML
